First cut flattening of parquet/file directory and consolidate writer, writer-internal headers
