+++ src/parquet/metadata-test.cc
@@ -15,8 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include <gtest/gtest.h>
 #include "parquet/file/metadata.h"
+#include <gtest/gtest.h>
 #include "parquet/schema.h"
 #include "parquet/statistics.h"
 #include "parquet/types.h"
